requirements.txt
blockchain_parser
mysql-connector-python